+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-01-27T12:37:49.956Z",
+    "2026-01-28T04:23:34.300Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {