+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -125,7 +125,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-01-27T12:37:53.424Z",
+    "2026-01-28T04:23:37.561Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {