+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -4099,7 +4099,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01KFZQC0J5WN24QZVZ7A76V77F"),
+  id: z.string().default("01KG1DFPZVX6MVF0VD4RM9W5XQ"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>